Add simple graphql integration test
Had problems with getting the custom DateTime scalar to work. Had to
provide it as bean for everything to work with the graphql template.
It is now being added in two places: GraphQlProvider and as a bean.
Need to next figure out how are the scalars added to the schema so
that I can remove either