Mock correct objects
GraphQLTest will fetch the resolvers by default. What we really need to
mock are the DAOs. This still doesnt resolve the fact that the scalars
arent fetched

Mock correct objects

GraphQLTest will fetch the resolvers by default. What we really need to
mock are the DAOs. This still doesnt resolve the fact that the scalars
arent fetched